+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
+
+## [0.3.2] - 2018-09-27
 ### Added
 * Support Heroku deployment
 * Allow user to specify `--enable-log` option
@@ -39,7 +41,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 * Docker automated build on Docker Hub
 * Support HTTPS
 
-[Unreleased]: https://github.com/nwtgck/piping-server/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/nwtgck/piping-server/compare/v0.3.2...HEAD
+[0.3.2]: https://github.com/nwtgck/piping-seraver/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/nwtgck/piping-seraver/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/nwtgck/piping-seraver/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/nwtgck/piping-seraver/compare/v0.2.0...v0.2.1